feat: add validation with pipes, start implement post request
