+++ src/users/users.controller.ts
@@ -1,4 +1,13 @@
-import { Controller, Get, Param, ParseUUIDPipe, Res } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Get,
+  Param,
+  ParseUUIDPipe,
+  Post,
+  Res,
+} from '@nestjs/common';
+import { CreateUserDto } from './dto/user.dto';
 import { Response } from 'express';
 import UsersService from './users.service';
 
@@ -18,6 +27,11 @@ export default class UsersController {
     @Res()
     res: Response,
   ) {
-    return this.usersService.getUserById(res, id);
+    return this.usersService.getUserById(id, res);
+  }
+
+  @Post()
+  createUser(@Body() body: CreateUserDto, @Res() res: Response) {
+    return this.usersService.createUser(body, res);
   }
 }
