+++ src/users/users.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import ResponseHelper from 'src/helpers/responseHelper';
 import db from 'src/data/inMemoryDB';
+import { CreateUserDto } from './dto/user.dto';
 import { Response } from 'express';
 
 @Injectable()
@@ -12,12 +13,16 @@ export default class UsersService {
     return users;
   }
 
-  getUserById(res: Response, id: string) {
+  getUserById(id: string, res: Response) {
     const user = db.user.find((user) => user.id === id);
     if (!user) {
       return ResponseHelper.sendNotFound(res, 'User not found');
     }
 
     return ResponseHelper.sendOk(res, user);
+  }
+
+  createUser(body: CreateUserDto, res: Response) {
+    const { login, password } = body;
   }
 }
